refactor testing.ts with default and used exec
